Revert "Merge branch 'refactorAudience' into feature/PhotonAttempt2"
This reverts commit 81aeb5f982a9d3882e9613577e037834e77d476f, reversing
changes made to 172cf533a14d6bd6a6cc44220c648d48b2539482.
