+++ TeamCode/src/main/java/edu/edina/opmodes/auto/Audience.java
@@ -3,7 +3,6 @@ package edu.edina.opmodes.auto;
 import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
-import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -24,10 +23,8 @@ public class Audience extends LinearOpMode {
     protected ActionManager manager;
     protected PoCMecanumDrive drive;
     protected PoCHuskyLens poCHuskyLens;
-
-    protected PropLocation propLocation = PropLocation.Right;
+    protected PropLocation propLocation = PropLocation.Center;
     protected PropLocation lastPropLocation = PropLocation.Right;
-    protected Vector2d backdropDropLocation;
 
     protected boolean makeSecondTrip = false;
     protected boolean yellowPixel = false;
@@ -239,10 +236,7 @@ public class Audience extends LinearOpMode {
 
             runPaths();
 
-            if (!makeSecondTrip) {
-                // Park
-                park(); // PK
-            }
+            park();
 
             hardware.stopCurrentMonitor();
         }
