+++ TeamCode/src/main/java/edu/edina/opmodes/auto/BlueAudienceCenterHC.java
@@ -1,6 +1,5 @@
-package edu.edina.opmodes.auto.state;
+package edu.edina.opmodes.auto;
 
-import com.acmerobotics.roadrunner.InstantAction;
 import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
@@ -11,29 +10,22 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 import edu.edina.library.enums.PropLocation;
-import edu.edina.library.util.RobotConfiguration;
-import edu.edina.opmodes.auto.BlueAudience;
 
 @Autonomous
-@Disabled
-public class BlueAudienceCenter extends BlueAudience {
+//@Disabled
+public class BlueAudienceCenterHC extends BlueAudience {
 
     private double stack1Y = 14.5;
-    private double stack2Y = 15;
+    private double stack2Y = 14.5;
 
     @Override
     protected void runPaths() {
+        int secondPickupHeight = -95;
+        double secondPickupX = -59.5;
+
         // If we want to drop Yellow..
         if (yellowPixel) {
-            if (propLocation == PropLocation.Left) {
-                Actions.runBlocking(
-                        drive.actionBuilder(drive.pose)
-                                // Head to Stacks
-                                .setReversed(true)
-                                .splineToSplineHeading(new Pose2d(-35, stack1Y, Math.toRadians(180)), Math.toRadians(90))
-                                .build()
-                );
-            } else if (propLocation == PropLocation.Center) {
+            if (propLocation == PropLocation.Center) {
                 // Drive to Stack Pick up 1st white
                 Actions.runBlocking(
                         drive.actionBuilder(drive.pose)
@@ -42,7 +34,7 @@ public class BlueAudienceCenter extends BlueAudience {
                                 .splineToSplineHeading(new Pose2d(-52, stack1Y, Math.toRadians(180)), Math.toRadians(90))
                                 .build()
                 );
-            } else {
+            } else if (propLocation == PropLocation.Right){
                 // Drive to Stack Pick up 1st white
                 Actions.runBlocking(
                         drive.actionBuilder(drive.pose)
@@ -53,76 +45,97 @@ public class BlueAudienceCenter extends BlueAudience {
                 );
             }
 
-            // Prepare lift, grab pixel, and raise lift
-            drive.turnBeamBreakOn(150);
+            if (propLocation == PropLocation.Right || propLocation == PropLocation.Center) {
+                // Prepare lift, grab pixel, and raise lift
+                drive.turnBeamBreakOn(150);
 
-            Actions.runBlocking(
-                    new SequentialAction(
-                            new ParallelAction(
-                                    manager.runLiftToPosition(-200),
-                                    manager.positionTheClawToPickupPixelsFromStack()
-                            ),
+                Actions.runBlocking(
+                        new SequentialAction(
+                                new ParallelAction(
+                                        manager.runLiftToPosition(-200),
+                                        manager.positionTheClawToPickupPixelsFromStack()
+                                ),
+                                drive.actionBuilder(drive.pose)
+                                        // Head to Stacks
+                                        .lineToX(-57)
+                                        .stopAndAdd(manager.closeLeftClaw())
+                                        .build()
+                        )
+                );
+
+                drive.turnBeamBreakOff();
+
+                // Check to see if delay is set and determine which routine to run
+                if (delayTime > 0) {  // Yes - delay is set
+                    Actions.runBlocking(
                             drive.actionBuilder(drive.pose)
-                                    // Head to Stacks
-                                    .lineToX(-57)
-                                    .stopAndAdd(manager.closeLeftClaw())
+                                    // Turn, then pack up and drive to backdrop
+                                    .lineToX(-55.5)
+                                    .stopAndAdd(manager.raiseLiftAfterStackPickup())
+                                    .lineToX(-50)
+                                    .afterDisp(0,
+                                            new ParallelAction(
+                                                    manager.lowerLiftForDriving(),
+                                                    manager.zeroLift(),
+                                                    manager.positionTheClawToDriveWithPixels()
+                                            )
+                                    )
+                                    .splineToSplineHeading(new Pose2d(new Vector2d(-30, 12), Math.toRadians(0)), Math.toRadians(0))
+                                    .splineToConstantHeading(new Vector2d(11, 12), Math.toRadians(0))
+                                    .waitSeconds(delayTime / 1000)  // Wait at midfield
+                                    .afterDisp(0, manager.getLiftReadyToDropThePixelHighOnTheWall())
+                                    .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
+                                    .lineToX(51.5)
+                                    .afterDisp(0, new SequentialAction(
+                                            manager.openRightClaw(),
+                                            new SleepAction(0.25),
+                                            manager.openLeftClaw()
+                                    ))
                                     .build()
-                    )
-            );
-
-            drive.turnBeamBreakOff();
-
-            // Check to see if delay is set and determine which routine to run
-            if (delayTime > 0) {  // Yes - delay is set
+                    );
+                } else {  // No delay - operate as normal
+                    Actions.runBlocking(
+                            drive.actionBuilder(drive.pose)
+                                    // Turn, then pack up and drive to backdrop
+                                    .lineToX(-56.5)
+                                    .stopAndAdd(manager.raiseLiftAfterStackPickup())
+                                    .lineToX(-50)
+                                    .afterDisp(0,
+                                            new ParallelAction(
+                                                    manager.lowerLiftForDriving(),
+                                                    manager.zeroLift(),
+                                                    manager.positionTheClawToDriveWithPixels()
+                                            )
+                                    )
+                                    .splineToSplineHeading(new Pose2d(new Vector2d(-30, 12), Math.toRadians(0)), Math.toRadians(0))
+                                    .splineToConstantHeading(new Vector2d(11, 12), Math.toRadians(0))
+                                    .afterDisp(0, manager.getLiftReadyToDropThePixelHighOnTheWall())
+                                    .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
+                                    .lineToX(51.5)
+                                    .afterDisp(0, new SequentialAction(
+                                            manager.openRightClaw(),
+                                            new SleepAction(0.25),
+                                            manager.openLeftClaw()
+                                    ))
+                                    .build()
+                    );
+                }
+            } else {
+                secondPickupHeight = -160;
+                stack2Y = 16;
+                secondPickupX = -63;
+                // just head to the backdrop to drop so we don't collide
                 Actions.runBlocking(
                         drive.actionBuilder(drive.pose)
-                                // Turn, then pack up and drive to backdrop
-                                .lineToX(-55.5)
-                                .stopAndAdd(manager.raiseLiftAfterStackPickup())
-                                .lineToX(-50)
-                                .afterDisp(0,
-                                        new ParallelAction(
-                                                manager.lowerLiftForDriving(),
-                                                manager.zeroLift(),
-                                                manager.positionTheClawToDriveWithPixels()
-                                        )
-                                )
-                                .splineToSplineHeading(new Pose2d(new Vector2d(-30, 12), Math.toRadians(0)), Math.toRadians(0))
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
-                                .splineToConstantHeading(new Vector2d(11, 12), Math.toRadians(0))
-                                .waitSeconds(delayTime/1000)  // Wait at midfield
+                                // Head to Stacks VIA C-Row
+                                .setReversed(true)
+                                .setTangent(Math.toRadians(200))
+                                .splineToSplineHeading(new Pose2d(new Vector2d(-35, 11), Math.toRadians(0)), Math.toRadians(0))
+                                .waitSeconds(1)
+                                .splineToSplineHeading(new Pose2d(new Vector2d(10, 11), Math.toRadians(0)), Math.toRadians(0))
                                 .afterDisp(0, manager.getLiftReadyToDropThePixelHighOnTheWall())
                                 .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
                                 .lineToX(51.5)
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
-                                .afterDisp(0, new SequentialAction(
-                                        manager.openRightClaw(),
-                                        new SleepAction(0.25),
-                                        manager.openLeftClaw()
-                                ))
-                                .build()
-                );
-            } else {  // No delay - operate as normal
-                Actions.runBlocking(
-                        drive.actionBuilder(drive.pose)
-                                // Turn, then pack up and drive to backdrop
-                                .lineToX(-56.5)
-                                .stopAndAdd(manager.raiseLiftAfterStackPickup())
-                                .lineToX(-50)
-                                .afterDisp(0,
-                                        new ParallelAction(
-                                                manager.lowerLiftForDriving(),
-                                                manager.zeroLift(),
-                                                manager.positionTheClawToDriveWithPixels()
-                                        )
-                                )
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
-                                .splineToSplineHeading(new Pose2d(new Vector2d(-30, 12), Math.toRadians(0)), Math.toRadians(0))
-                                .splineToConstantHeading(new Vector2d(11, 12), Math.toRadians(0))
-                                .afterDisp(0, manager.getLiftReadyToDropThePixelLowOnTheWallAsync())
-                                .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
-                                .lineToX(51.5)
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
                                 .afterDisp(0, new SequentialAction(
                                         manager.openRightClaw(),
                                         new SleepAction(0.25),
@@ -152,10 +165,6 @@ public class BlueAudienceCenter extends BlueAudience {
         // If we are making a second trip to the stacks
         if (makeSecondTrip) {
             // go get other white pixels
-            if ((propLocation == PropLocation.Right || propLocation == PropLocation.Center) && dropOnBackdrop) {
-                dropOnBackdrop = false;
-                dropOnBackstage = true;
-            }
 
             Actions.runBlocking(
                     new ParallelAction(
@@ -165,10 +174,8 @@ public class BlueAudienceCenter extends BlueAudience {
                             ),
                             drive.actionBuilder(drive.pose)
                                     .lineToX(46)
-                                    .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
                                     .splineToSplineHeading(new Pose2d(11, stack2Y, Math.toRadians(180)), Math.toRadians(180))
                                     .splineToConstantHeading(new Vector2d(-48, stack2Y), Math.toRadians(180))
-                                    .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
                                     .build()
                     )
             );
@@ -180,11 +187,11 @@ public class BlueAudienceCenter extends BlueAudience {
                     new SequentialAction(
                             new ParallelAction(
                                     manager.positionTheClawToPickupPixelsFromStack(),
-                                    manager.runLiftToPosition(-95)
+                                    manager.runLiftToPosition(secondPickupHeight)
                             ),
                             drive.actionBuilder(drive.pose)
                                     // Head to Stacks
-                                    .lineToX(-59.5)
+                                    .lineToX(secondPickupX)
                                     .build(),
                             new ParallelAction(
                                     manager.closeLeftClaw(),
@@ -203,7 +210,6 @@ public class BlueAudienceCenter extends BlueAudience {
                         drive.actionBuilder(drive.pose)
                                 // Head to Stacks VIA A-Row
                                 .lineToX(-50)
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
                                 .afterDisp(0,
                                         new ParallelAction(
                                                 manager.lowerLiftForDriving(),
@@ -215,7 +221,6 @@ public class BlueAudienceCenter extends BlueAudience {
                                 .afterDisp(2, manager.getLiftReadyToDropPixelFromRight())
                                 .splineTo(new Vector2d(30, 13), Math.toRadians(0))
                                 .splineTo(new Vector2d(50, 21), Math.toRadians(35))
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
                                 .stopAndAdd(manager.openLeftClaw())
                                 .afterTime(0.1, manager.openAutoClaw())
                                 .waitSeconds(0.25)
@@ -238,7 +243,6 @@ public class BlueAudienceCenter extends BlueAudience {
                         drive.actionBuilder(drive.pose)
                                 // Head to Stacks VIA A-Row
                                 .lineToX(-48)
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
                                 .afterDisp(0,
                                         new ParallelAction(
                                                 manager.lowerLiftForDriving(),
@@ -247,8 +251,7 @@ public class BlueAudienceCenter extends BlueAudience {
                                         ))
                                 .setReversed(true)
                                 .splineToSplineHeading(new Pose2d(new Vector2d(-35, 12), Math.toRadians(0)), Math.toRadians(0))
-                                .splineToConstantHeading(new Vector2d(56, 14), Math.toRadians(0))
-                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
+                                .splineToConstantHeading(new Vector2d(56, 12), Math.toRadians(0))
                                 .afterDisp(0, new SequentialAction(
                                         manager.openAutoClaw(),
                                         manager.openLeftClaw()
