+++ TeamCode/src/main/java/edu/edina/opmodes/auto/RedAudienceCenter.java
@@ -1,264 +1,560 @@
 package edu.edina.opmodes.auto;
 
 import com.acmerobotics.roadrunner.InstantAction;
+import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
+import com.acmerobotics.roadrunner.SleepAction;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
+import com.qualcomm.hardware.rev.RevBlinkinLedDriver;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
+import org.firstinspires.ftc.teamcode.PoCMecanumDrive;
+
+import edu.edina.library.actions.roadrunner.ActionManager;
+import edu.edina.library.enums.Alliance;
+import edu.edina.library.enums.ParkLocation;
 import edu.edina.library.enums.PropLocation;
+import edu.edina.library.util.PoCHuskyLens;
+import edu.edina.library.util.RobotConfiguration;
+import edu.edina.library.util.RobotHardware;
+import edu.edina.library.util.SmartGamepad;
 
 @Autonomous
 //@Disabled
-public class RedAudienceCenter extends RedAudience {
+public class RedAudienceCenter extends LinearOpMode {
+    protected RobotHardware hardware;
+    protected ActionManager manager;
+    protected PoCMecanumDrive drive;
+    protected PoCHuskyLens poCHuskyLens;
+    protected PropLocation propLocation = PropLocation.Center;
+
+    private boolean makeSecondTrip = false;
+    private boolean yellowPixel = false;
+    private boolean dropOnBackdrop = false;
+    private boolean dropOnBackstage = false;
+    private boolean useCamera = true;
+
+    private ParkLocation parkLocation = ParkLocation.None;
+
+    private long delayTime = 0;
+
+    protected void initHardware() {
+        hardware = new RobotHardware(hardwareMap);
+        manager = new ActionManager(hardware);
+
+        drive = new PoCMecanumDrive(hardware.leftFront,
+                hardware.leftBack, hardware.rightBack, hardware.rightFront,
+                hardware.par0, hardware.perp,
+                hardware.externalImu, hardware.expansionImu, hardware.voltageSensor, hardware.beamBreak, getStartPose());
+
+        // HuskyLens Init
+        poCHuskyLens = new PoCHuskyLens(hardware.huskyLens, telemetry, getAlliance());
+        poCHuskyLens.init();
+
+        hardware.dropServosForAutonomous();
+        hardware.droneLaunchServo.setPosition(RobotConfiguration.getInstance().droneLauncherArmedPosition);
+
+        manager.init();
+        manager.start();
+    }
+
+    protected Alliance getAlliance() {
+        return Alliance.Red;
+    }
+
+    protected RevBlinkinLedDriver.BlinkinPattern getUnsuccessfulPropMatchColor() {
+        return RevBlinkinLedDriver.BlinkinPattern.GREEN;
+    }
+
+    protected RevBlinkinLedDriver.BlinkinPattern getSuccessfulPropMatchColor() {
+        return RevBlinkinLedDriver.BlinkinPattern.HEARTBEAT_WHITE;
+    }
+
+    protected Pose2d getStartPose() {
+        return new Pose2d(-42, -64, Math.toRadians(90));
+    }
 
     @Override
-    protected void runPaths() {
+    public void runOpMode() throws InterruptedException {
+        SmartGamepad pad1 = new SmartGamepad(gamepad1);
 
-        if (yellowPixel) {  // P + Y + 1W path
-            purpleToStack(); // B
+        initHardware();
 
-            // If we are doing anything but stopping after purple
-            if (makeSecondTrip) {  // P + Y + 3W path
-                stackToYellowToStack(); // DPAD_UP/DOWN
-                if (dropOnBackdrop) {
-                    stackToAngleDrop(); // DPAD_UP
-                } else if (dropOnBackstage) {
-                    stackToBackStageDrop(); // DPAD_DOWN
+        Actions.runBlocking(manager.positionTheClawToPickupPixels());
+
+        // Turn on prop illumination
+        hardware.lights.setPower(1);
+
+        while (!isStarted()) {
+            pad1.update();
+
+            telemetry.addData("A for P only", "");
+            telemetry.addData("X for P, Y, 1W and park in corner", "");
+            telemetry.addData("Y for P, Y, 1W and park in center", "");
+            telemetry.addData("DPAD-UP for P, Y, 3Ws on backdrop park in front", "");
+            telemetry.addData("DPAD-DN for P, Y, 3Ws and park in center", "");
+            telemetry.addData("L-BUMPER to increase delay, R-BUMPER to decrease delay.", "");
+            telemetry.addData("L-TRIGGER to close claws, L-TRIGGER to open", "");
+            telemetry.addData("LEFT-STICK-DOWN : manual rotate prop position", "");
+            telemetry.addData("RIGHT-STICK-DOWN : manual or auto camera", "");
+
+            if (pad1.a) {
+                yellowPixel = false;
+                makeSecondTrip = false;
+                dropOnBackdrop = false;
+                dropOnBackstage = false;
+                parkLocation = ParkLocation.None;
+            }
+
+            if (pad1.x) {
+                parkLocation = ParkLocation.Corner;
+                yellowPixel = true;
+                makeSecondTrip = false;
+                dropOnBackdrop = false;
+                dropOnBackstage = false;
+            }
+
+            if (pad1.y) {
+                parkLocation = ParkLocation.Center;
+                yellowPixel = true;
+                makeSecondTrip = false;
+                dropOnBackdrop = false;
+                dropOnBackstage = false;
+            }
+
+            if (pad1.dpad_up) {
+                yellowPixel = true;
+                makeSecondTrip = true;
+                dropOnBackdrop = true;
+                dropOnBackstage = false;
+                parkLocation = ParkLocation.None;
+            }
+
+            if (pad1.dpad_down) {
+                yellowPixel = true;
+                makeSecondTrip = true;
+                dropOnBackdrop = false;
+                dropOnBackstage = true;
+                parkLocation = ParkLocation.None;
+            }
+
+            // Delay - Max of 4000ms, Min of 0ms
+            if (pad1.left_bumper) {
+                delayTime += (delayTime > 3000) ? 0 : 1000;
+            } else if (pad1.right_bumper) {
+                delayTime -= (delayTime < 1000) ? 0 : 1000;
+            }
+
+            // If we have ANY delay, don't allow second trip
+            makeSecondTrip = (delayTime > 0) ? false : makeSecondTrip;
+
+            // Close the claws
+            if (gamepad1.left_trigger != 0) {
+                Actions.runBlocking(
+                        new SequentialAction(
+                                new ParallelAction(
+                                        manager.closeRightClaw(),
+                                        manager.closeLeftClaw(),
+                                        manager.openAutoClaw()
+                                ),
+                                manager.positionTheClawToDriveWithPixels())
+                );
+            }
+
+            // Open the claws
+            if (gamepad1.right_trigger != 0) {
+                Actions.runBlocking(new ParallelAction(
+                        manager.openRightClaw(),
+                        manager.openLeftClaw(),
+                        manager.openAutoClaw(),
+                        manager.positionTheClawToPickupPixels()
+                ));
+            }
+
+            // Select whether camera is live or not
+            if (pad1.right_stick_button) {
+                if (useCamera) {
+                    useCamera = false;
+                } else {
+                    useCamera = true;
+                }
+            }
+
+            // Find Prop Location
+            if (useCamera) {
+                // Find Prop Location
+                poCHuskyLens.update();
+                propLocation = poCHuskyLens.getPropLocation();
+                if (propLocation == PropLocation.None) {
+                    propLocation = PropLocation.Right;
                 }
             } else {
-                stackToYellow(); // C
+                if (pad1.left_stick_button) {
+                    switch (propLocation) {
+                        case None:
+                            propLocation = PropLocation.Left;
+                            break;
+                        case Left:
+                            propLocation = PropLocation.Center;
+                            break;
+                        case Center:
+                            propLocation = PropLocation.Right;
+                            break;
+                        case Right:
+                            propLocation = PropLocation.Left;
+                            break;
+                        case Idle:
+                            propLocation = PropLocation.Left;
+                            break;
+                        default:
+                            propLocation = PropLocation.Center;
+                            break;
+                    }
+                }
+            }
+
+            telemetry.addData("==========================", "");
+            telemetry.addData("Drop Yellow Pixel", yellowPixel);
+            telemetry.addData("Make Second Trip", makeSecondTrip);
+            telemetry.addData("Current Park Location", parkLocation);
+            telemetry.addData("Drop on backdrop", dropOnBackdrop);
+            telemetry.addData("Drop on center", dropOnBackstage);
+            telemetry.addData("Delay in seconds", delayTime / 1000);
+            telemetry.addData("Location", propLocation);
+            telemetry.addData("Use Camera", useCamera);
+            telemetry.update();
+        }
+
+        // Turn off prop lighting
+        hardware.lights.setPower(0);
+
+        if (opModeIsActive()) {
+
+            hardware.startCurrentMonitor();
+
+            runPaths();
+
+            hardware.stopCurrentMonitor();
+        }
+    }
+
+    protected void runPaths() {
+        Vector2d propDropLocation;
+        Vector2d backdropDropLocation;
+
+        // Determine location for purple pixel
+        switch(propLocation) {
+            case Left:
+                propDropLocation = new Vector2d(-40, -38);
+                backdropDropLocation = new Vector2d(49,-32);
+                break;
+            case Right:
+                propDropLocation = new Vector2d(-30, -38);
+                backdropDropLocation = new Vector2d(49.5,-46);
+                break;
+            case Center:
+            default:
+                propDropLocation = new Vector2d(-31, -33);
+                backdropDropLocation = new Vector2d(49,-38);
+                break;
+        }
+
+        // Run to drop PURPLE pixel
+        switch (propLocation) {
+            case Right:
+                Actions.runBlocking(
+                        new SequentialAction(
+                                drive.actionBuilder(drive.pose)
+                                        .splineTo(propDropLocation, Math.toRadians(45))
+                                        .build(),
+                                manager.openLeftClaw()
+                        )
+                );
+                break;
+            case Left:
+                Actions.runBlocking(
+                        new SequentialAction(
+                                drive.actionBuilder(drive.pose)
+                                        .splineTo(propDropLocation, Math.toRadians(135))
+                                        .build(),
+                                manager.openLeftClaw()
+                        )
+                );
+                break;
+            default:
+                // Execute drive to prop drop spot and drop
+                Actions.runBlocking(
+                        new SequentialAction(
+                                drive.actionBuilder(drive.pose)
+                                        .splineTo(propDropLocation, Math.toRadians(90))
+                                        .build(),
+                                manager.openLeftClaw()
+                        )
+                );
+                break;
+        }
+
+        // If we want to drop Yellow..
+        if (yellowPixel) {
+            if (propLocation == PropLocation.Left) {
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(-35, -12.5, Math.toRadians(180)), Math.toRadians(45))
+                                .build()
+                );
+            } else if (propLocation == PropLocation.Center) {
+                // Drive to Stack Pick up 1st white
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(-48, -18, Math.toRadians(180)), Math.toRadians(90))
+                                .splineToSplineHeading(new Pose2d(-48, -10, Math.toRadians(180)), Math.toRadians(90))
+                                .build()
+                );
+            } else {
+                // Drive to Stack Pick up 1st white
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(-48, -10, Math.toRadians(180)), Math.toRadians(90))
+                                .build()
+                );
+            }
+
+            drive.turnBeamBreakOn();
+
+            Actions.runBlocking(
+                    new SequentialAction(
+                            new ParallelAction(
+                                    manager.runLiftToPosition(-200),
+                                    manager.positionTheClawToPickupPixelsFromStack()
+                            ),
+                            drive.actionBuilder(drive.pose)
+                                    // Head to Stacks
+                                    .lineToX(-58)
+                                    .stopAndAdd(manager.closeLeftClaw())
+                                    .build()
+                    )
+            );
+
+            drive.turnBeamBreakOff();
+
+            // Check to see if there is delay - if so, run special version with wait during return
+            if (delayTime > 0) {  // Yes, there's a delay
+                // drive to backstage - 1st trip with delay at center field
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks VIA C-Row
+                                .lineToX(-57.5)
+                                .stopAndAdd(manager.raiseLiftAfterStackPickup())
+                                .lineToX(-48)
+                                .afterDisp(0,
+                                        new ParallelAction(
+                                                manager.lowerLiftForDriving(),
+                                                manager.zeroLift(),
+                                                manager.positionTheClawToDriveWithPixels()
+                                        ))
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(new Vector2d(-35, -11), Math.toRadians(0)), Math.toRadians(0))
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
+                                .splineTo(new Vector2d(10, -11), Math.toRadians(0))
+                                .waitSeconds(delayTime/1000)
+                                .afterDisp(0, manager.getLiftReadyToDropThePixelHighOnTheWall())
+                                .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
+                                .lineToX(51.5)
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
+                                .afterDisp(0, new SequentialAction(
+                                        manager.openRightClaw(),
+                                        new SleepAction(0.25),
+                                        manager.openLeftClaw()
+                                ))
+                                .build()
+                );
+            } else {
+                // drive to backstage - 1st trip - no delay
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks VIA C-Row
+                                .lineToX(-57.5)
+                                .stopAndAdd(manager.raiseLiftAfterStackPickup())
+                                .lineToX(-48)
+                                .afterDisp(0,
+                                        new ParallelAction(
+                                                manager.lowerLiftForDriving(),
+                                                manager.zeroLift(),
+                                                manager.positionTheClawToDriveWithPixels()
+                                        ))
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(new Vector2d(-35, -11), Math.toRadians(0)), Math.toRadians(0))
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
+                                .afterDisp(25, manager.getLiftReadyToDropThePixelHighOnTheWall())
+                                .splineToConstantHeading(new Vector2d(10, -11), Math.toRadians(0))
+                                .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
+                                .lineToX(51.5)
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
+                                .afterDisp(0, new SequentialAction(
+                                        manager.openRightClaw(),
+                                        new SleepAction(0.25),
+                                        manager.openLeftClaw()
+                                ))
+                                .build()
+                );
+            }
+
+            // If we're done making stack trips
+            if (!makeSecondTrip) {
+                // back away and pack up
+                Actions.runBlocking(
+                        new ParallelAction(
+                                drive.actionBuilder(drive.pose)
+                                        .lineToX(44)
+                                        .build(),
+                                new SequentialAction(
+                                        new SleepAction(.2),
+                                        manager.getLiftReadyToDrive()
+                                )
+                        )
+                );
             }
         }
-    }
 
-    protected void purpleToStack() {
+        // If we are making a second trip to the stacks
+        if (makeSecondTrip) {
+            // go get other white pixels
+            if ((propLocation == PropLocation.Left || propLocation == PropLocation.Center) && dropOnBackdrop) {
+                dropOnBackdrop = false;
+                dropOnBackstage = true;
+            }
 
-        if (propLocation == PropLocation.Left) {
             Actions.runBlocking(
-                    drive.actionBuilder(drive.pose)
-                            // Head to Stacks
-                            .setReversed(true)
-                            .splineToSplineHeading(new Pose2d(-35, -12, Math.toRadians(180)), Math.toRadians(45))
-                            // Prepare for grabbing - Trip 1
-                            .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOn(150)))
-                            .afterDisp(0, manager.runLiftToPosition(-180))
-                            .afterDisp(0, manager.positionTheClawToPickupPixelsFromStack())
-                            .setReversed(false)
-                            .splineToSplineHeading(new Pose2d(-55, -12, Math.toRadians(180)), Math.toRadians(180))
-                            .lineToX(-58)
-                            .build()
+                    new ParallelAction(
+                            new SequentialAction(
+                                    new SleepAction(.2),
+                                    manager.getLiftReadyToDrive()
+                            ),
+                            drive.actionBuilder(drive.pose)
+                                    .lineToX(44)
+                                    .setReversed(true)
+                                    .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
+                                    .splineToSplineHeading(new Pose2d(10, -12.5, Math.toRadians(-180)), Math.toRadians(180))
+                                    .splineToConstantHeading(new Vector2d(-54, -12.5), Math.toRadians(180))
+                                    .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
+                                    .build()
+                    )
             );
-        } else if (propLocation == PropLocation.Center) {
-            // Drive to Stack Pick up 1st white
+
+            // Reach out, grab pixels, close the claws
+            drive.turnBeamBreakOn();
+
             Actions.runBlocking(
-                    drive.actionBuilder(drive.pose)
-                            // Head to Stacks
-                            .setReversed(true)
-                            .splineToSplineHeading(new Pose2d(-48, -18, Math.toRadians(180)), Math.toRadians(90))
-                            .splineToSplineHeading(new Pose2d(-50, -12, Math.toRadians(180)), Math.toRadians(90))
-                            .setReversed(false)
-                            // Prepare for grabbing - Trip 1
-                            .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOn(150)))
-                            .afterDisp(0, manager.runLiftToPosition(-180))
-                            .afterDisp(0, manager.positionTheClawToPickupPixelsFromStack())
-                            .lineToX(-58)
-                            .build()
+                    new SequentialAction(
+                            new ParallelAction(
+                                    manager.positionTheClawToPickupPixelsFromStack(),
+                                    manager.runLiftToPosition(-125)
+                            ),
+                            drive.actionBuilder(drive.pose)
+                                    // Head to Stacks
+                                    .lineToX(-62)
+                                    .build(),
+                            new ParallelAction(
+                                    manager.closeLeftClaw(),
+                                    manager.closeAutoClaw()
+                            ),
+                            new SleepAction(0.2)
+                    )
             );
-        } else {
-            // Drive to Stack Pick up 1st white
-            Actions.runBlocking(
-                    drive.actionBuilder(drive.pose)
-                            // Head to Stacks
-                            .setReversed(true)
-                            .splineToSplineHeading(new Pose2d(-50, -12, Math.toRadians(180)), Math.toRadians(90))
-                            // Prepare for grabbing - Trip 1
-                            .setReversed(false)
-                            .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOn(150)))
-                            .afterDisp(0, manager.runLiftToPosition(-180))
-                            .afterDisp(0, manager.positionTheClawToPickupPixelsFromStack())
-                            .lineToX(-58)
-                            .build()
-            );
+
+            drive.turnBeamBreakOff();
+
+            // If we're going to drop on the background
+            if (dropOnBackdrop) {
+                // drive to backstage - 2nd trip
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks VIA A-Row
+                                .lineToX(-48)
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
+                                .afterDisp(0,
+                                        new ParallelAction(
+                                                manager.lowerLiftForDriving(),
+                                                manager.zeroLift(),
+                                                manager.positionTheClawToDriveWithPixels()
+                                        ))
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(new Vector2d(-35, -11), Math.toRadians(0)), Math.toRadians(0))
+                                .splineTo(new Vector2d(10, -11), Math.toRadians(0))
+                                .afterDisp(0, manager.getLiftReadyToDropPixelFromLeft())
+                                .splineTo(new Vector2d(48, -19), Math.toRadians(-25))
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOff()))
+                                .stopAndAdd(manager.openLeftClaw())
+                                .afterTime(0.1, manager.openAutoClaw())
+                                .waitSeconds(0.25)
+                                .build());
+
+                // back away and pack up
+                Actions.runBlocking(
+                        new SequentialAction(
+                                drive.actionBuilder(drive.pose)
+                                        .lineToX(43)
+                                        .build(),
+                                manager.getLiftReadyToDrive()
+                        )
+                );
+            }
+
+            // If we are dropping on Backstage
+            if (dropOnBackstage) {
+                Actions.runBlocking(
+                        drive.actionBuilder(drive.pose)
+                                // Head to Stacks VIA A-Row
+                                .lineToX(-48)
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
+                                .afterDisp(0,
+                                        new ParallelAction(
+                                                manager.lowerLiftForDriving(),
+                                                manager.zeroLift(),
+                                                manager.positionTheClawToDriveWithPixels()
+                                        ))
+                                .setReversed(true)
+                                .splineToSplineHeading(new Pose2d(new Vector2d(-35, -11), Math.toRadians(0)), Math.toRadians(0))
+                                .splineTo(new Vector2d(10, -11), Math.toRadians(0))
+                                .splineTo(new Vector2d(48, -12), Math.toRadians(0))
+                                .afterDisp(0, new InstantAction(() -> drive.turnErrorPoseStopOn()))
+                                .stopAndAdd(manager.openLeftClaw())
+                                .afterTime(0.1, manager.openAutoClaw())
+                                .waitSeconds(0.25)
+                                .lineToX(44)
+                                .build());
+
+            }
         }
-    }
 
-    protected void stackToYellow() {
-        // Run to backdrop.
-        Actions.runBlocking(
-                drive.actionBuilder(drive.pose)
-                        // Back away and pack up
-                        .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOff()))
-
-                        // Move in and grab pixels until beam break
-                        .afterTime(0, manager.closeLeftClaw())
-
-                        // Back away a little and raise the lift
-                        .lineToX(-56.5)
-                        .stopAndAdd(manager.raiseLiftAfterStackPickup())
-                        .waitSeconds(delayTime/1000) // Optional Wait
-
-                        // Finish backing away and prepare to drive
-                        .lineToX(-53)
-                        .afterDisp(3, manager.lowerLiftForDriving())
-                        .afterDisp(3, manager.zeroLift())
-                        .afterDisp(3, manager.positionTheClawToDriveWithPixels())
-
-                        // Head to backdrop
-                        .setReversed(true)
-                        .splineToSplineHeading(new Pose2d(new Vector2d(-35, -11), Math.toRadians(0)), Math.toRadians(0))
-                        .afterDisp(25, manager.getLiftReadyToDropThePixelHighOnTheWall())
-                        .splineToConstantHeading(new Vector2d(10, -11), Math.toRadians(0))
-                        .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
-
-                        // Release Yellow + White
-                        .stopAndAdd(manager.openRightClaw())
-                        .afterTime(0.25, manager.openLeftClaw())
-                        .waitSeconds(0.25)
-                        .build()
-        );
-
-    }
-
-    protected void stackToYellowToStack() {
-
-        // Run to backdrop.
-        Actions.runBlocking(
-                drive.actionBuilder(drive.pose)
-                        // Disable beam break
-                        .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOff()))
-
-                        // Grab Pixels
-                        .afterTime(0, manager.closeAutoClaw())
-                        .afterTime(0, manager.closeLeftClaw())
-                        .afterTime(0, manager.closeRightClaw())
-
-                        // Back away a little and raise the lift
-                        .lineToX(-56.5)
-                        .stopAndAdd(manager.raiseLiftAfterStackPickup())
-                        .waitSeconds(delayTime/1000)
-
-                        // Finish backing away and prepare to drive
-                        .lineToX(-53)
-                        .afterDisp(3, manager.lowerLiftForDriving())
-                        .afterDisp(3, manager.zeroLift())
-                        .afterDisp(3, manager.positionTheClawToDriveWithPixels())
-
-                        // Return to backdrop
-                        .setReversed(true)
-                        .splineToSplineHeading(new Pose2d(new Vector2d(-35, -11), Math.toRadians(0)), Math.toRadians(0))
-                        .afterDisp(25, manager.getLiftReadyToDropThePixelHighOnTheWall())
-                        .splineToConstantHeading(new Vector2d(10, -11), Math.toRadians(0))
-                        .splineToConstantHeading(backdropDropLocation, Math.toRadians(0))
-
-                        // Release Yellow + White
-                        .stopAndAdd(manager.openRightClaw())
-                        .afterTime(0.25, manager.openLeftClaw())
-                        .waitSeconds(0.25)
-
-                        // Back up and pack up
-                        .lineToX(50)
-                        .afterDisp(0, manager.getLiftReadyToDrive())
-
-                        // Drive back to stacks
-                        .setReversed(true)
-                        .splineToSplineHeading(new Pose2d(24, -10.5, Math.toRadians(180)), Math.toRadians(180))
-                        .splineTo(new Vector2d(-44, -10.5), Math.toRadians(180))
-
-                        // Prepare for grabbing - Trip 2
-                        .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOn(150)))
-                        .afterDisp(0, manager.runLiftToPosition(-123))
-                        .afterDisp(0, manager.positionTheClawToPickupPixelsFromStack())
-                        .lineToX(-53)
-                        .build()
-        );
-    }
-
-    protected void stackToAngleDrop () {
-        Actions.runBlocking(
-                drive.actionBuilder(drive.pose)
-                        // Turn off Beam Break
-                        .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOff()))
-
-                        // Grab Pixels
-                        .afterTime(0, manager.closeAutoClaw())
-                        .afterTime(0, manager.closeLeftClaw())
-                        .afterTime(0, manager.closeRightClaw())
-
-                        // Back away a little and raise the lift
-                        .waitSeconds(0.1)
-                        .stopAndAdd(manager.raiseLiftAfterStackPickup())
-
-                        // Finish backing away and prepare to drive
-                        .afterDisp(3, manager.lowerLiftForDriving())
-                        .afterDisp(3, manager.zeroLift())
-                        .afterDisp(3, manager.positionTheClawToDriveWithPixels())
-                        .lineToX(-50)
-
-                        // Return to backdrop and angle drop
-                        .setReversed(true)
-                        .splineToSplineHeading(new Pose2d(-11, -12, Math.toRadians(0)), Math.toRadians(0))
-                        .afterDisp(30, manager.getLiftReadyToDropPixelFromLeft())
-                        .splineTo(new Vector2d(40, -12), Math.toRadians(0))
-                        .splineTo(new Vector2d(61, -20), Math.toRadians(-35))
-
-                        // Release all pixels
-                        .afterTime(0, manager.openAutoClaw())
-                        .afterTime(0, manager.openLeftClaw())
-                        .afterTime(0, manager.openRightClaw())
-                        .waitSeconds(0.25)
-                        .build()
-        );
-
-        Actions.runBlocking(new SequentialAction(
-                drive.actionBuilder(drive.pose)
-                        // Back up and pack up
-                        .lineToX(58)
-                        .afterDisp(1, manager.getLiftReadyToDrive())
-                        .build()));
-
-    }
-
-    protected void stackToBackStageDrop() {
-        Actions.runBlocking(
-                drive.actionBuilder(drive.pose)
-                        // Turn off Beam Break
-                        .afterTime(0, new InstantAction(() -> drive.turnBeamBreakOff()))
-
-                        // Grab Pixels
-                        .afterTime(0, manager.closeAutoClaw())
-                        .afterTime(0, manager.closeLeftClaw())
-                        .afterTime(0, manager.closeRightClaw())
-
-                        // Back away a little and raise the lift
-                        .waitSeconds(0.1)
-                        .stopAndAdd(manager.raiseLiftAfterStackPickup())
-
-                        // Finish backing away and prepare to drive
-                        .afterDisp(3, manager.lowerLiftForDriving())
-                        .afterDisp(3, manager.zeroLift())
-                        .afterDisp(3, manager.positionTheClawToDriveWithPixels())
-                        .lineToX(-50)
-
-                        // Return to backdrop and angle drop
-                        .setReversed(true)
-                        .splineToSplineHeading(new Pose2d(-11, -12, Math.toRadians(0)), Math.toRadians(0))
-                        .splineTo(new Vector2d(61, -12), Math.toRadians(0))
-
-                        // Release all pixels
-                        .afterTime(0, manager.openAutoClaw())
-                        .afterTime(0, manager.openLeftClaw())
-                        .afterTime(0, manager.openRightClaw())
-                        .waitSeconds(0.25)
-                        .build()
-        );
-
-        Actions.runBlocking(new SequentialAction(
-                drive.actionBuilder(drive.pose)
-                        // Back up
-                        .lineToX(58)
-                        .build()));
-
+        // park
+        switch (parkLocation) {
+            case Center:
+                Actions.runBlocking(new SequentialAction(
+                        drive.actionBuilder(drive.pose)
+                                .setReversed(true)
+                                .splineTo(new Vector2d(58, -14), Math.toRadians(0))
+                                .build()));
+                break;
+            case Corner:
+                Actions.runBlocking(new SequentialAction(
+                        drive.actionBuilder(drive.pose)
+                                .setReversed(true)
+                                .splineTo(new Vector2d(58, -60), Math.toRadians(0))
+                                .build()));
+                break;
+            default:
+                break;
+        }
     }
 }
